Implemented skeleton of TIPLocation object. Fixed some issues with our server objects. Corrected distance calculation so we aren't off by 1.

+++ server/src/main/java/com/tripco/t18/misc/GreatCircleDistance.java
@@ -8,30 +8,30 @@ import java.util.Map;
 /** Determines the distance between geographic coordinates.
  */
 public class GreatCircleDistance {
-  public int calculateDistance(Map origin, Map destination, float earthRadius) {
-    double origin_lat = to_radians(Double.parseDouble((String) origin.get("latitude")));
-    double origin_lon = to_radians(Double.parseDouble((String) origin.get("longitude")));
+  public int calculateDistance(Map origin, Map destination, Double earthRadius) {
+    Double origin_lat = to_radians(Double.parseDouble((String) origin.get("latitude")));
+    Double origin_lon = to_radians(Double.parseDouble((String) origin.get("longitude")));
 
-    double dest_lat = to_radians(Double.parseDouble((String) destination.get("latitude")));
-    double dest_lon = to_radians(Double.parseDouble((String) destination.get("longitude")));
-    double delta_lon = Math.abs(dest_lon - origin_lon);
+    Double dest_lat = to_radians(Double.parseDouble((String) destination.get("latitude")));
+    Double dest_lon = to_radians(Double.parseDouble((String) destination.get("longitude")));
+    Double delta_lon = Math.abs(dest_lon - origin_lon);
 
-    double numerator = Math.pow(Math.cos(dest_lat) * Math.sin(delta_lon), 2);
+    Double numerator = Math.pow(Math.cos(dest_lat) * Math.sin(delta_lon), 2);
     numerator += Math.pow(
         Math.cos(origin_lat) * Math.sin(dest_lat) - Math.sin(origin_lat) * Math.cos(dest_lat) * Math
             .cos(delta_lon), 2);
     numerator = Math.sqrt(numerator);
 
-    double denominator =
+    Double denominator =
         Math.sin(origin_lat) * Math.sin(dest_lat) + Math.cos(origin_lat) * Math.cos(dest_lat) * Math
             .cos(delta_lon);
-    double arc = Math.atan2(numerator,denominator);
+    Double arc = Math.atan2(numerator,denominator);
 
-    return Math.round((int) (arc * earthRadius));
+    return (int) Math.round((arc * earthRadius));
   }
 
-  public double to_radians(double degrees) {
-    double pi = Math.PI;
+  public Double to_radians(Double degrees) {
+    Double pi = Math.PI;
     return degrees * (pi/180);
   }
 }
